samples/bot-quickstart/python/bot-quickstart/pyproject.toml
[project]
name = "bot-quickstart"
version = "0.1.0"
description = "This sample demonstrates a Bot Quickstart for Microsoft Teams using Python."
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "python-dotenv>=1.0.*",
    "microsoft-teams-apps>=2.0.*",
]
